+++ com.unity.cinemachine/Runtime/Core/SplineContainerExtensions.cs
@@ -14,20 +14,20 @@ namespace Unity.Cinemachine
         public static bool IsValid(this ISplineContainer spline) => spline != null && spline.Splines != null && spline.Splines.Count > 0;
 
         /// <summary>
-        /// Apply to a <see cref="SplineContainer"/>additional roll from <see cref="CinemachineSplineRoll"/>
+        /// Apply to a <see cref="CachedScaledSpline"/>additional roll from <see cref="CinemachineSplineRoll"/>
         /// </summary>
         /// <param name="spline">The spline in question</param>
         /// <param name="roll">The additional roll to apply</param>
         /// <param name="tNormalized">The normalized position on the spline</param>
-        /// <param name="getLength">Delegate to get the path length including scale</param>
+        /// <param name="defaultRotation">Used to resolve cases where spline tangent is undefined</param>
         /// <param name="position">returned point on the spline, in spline-local coords</param>
         /// <param name="rotation">returned rotation at the point on the spline, in spline-local coords</param>
         /// <returns>True if the spline position is valid</returns>
         public static bool LocalEvaluateSplineWithRoll(
-            this CachedScaledSpline spline,
-            CinemachineSplineRoll roll,
-            Quaternion defaultRotation,
+            this ISpline spline,
             float tNormalized, 
+            Quaternion defaultRotation,
+            CinemachineSplineRoll roll,
             out Vector3 position, out Quaternion rotation)
         {
             if (spline == null || !SplineUtility.Evaluate(
@@ -72,29 +72,46 @@ namespace Unity.Cinemachine
         }
 
         /// <summary>
-        /// Apply to a <see cref="SplineContainer"/>additional roll from <see cref="CinemachineSplineRoll"/>
+        /// Apply to a <see cref="CachedScaledSpline"/>additional roll from <see cref="CinemachineSplineRoll"/>
         /// </summary>
         /// <param name="spline">The spline in question</param>
-        /// <param name="roll">The additional roll to apply</param>
         /// <param name="tNormalized">The normalized position on the spline</param>
-        /// <param name="getLength">Delegate to get the path length including scale</param>
+        /// <param name="defaultRotation">Used to resolve cases where spline tangent is undefined</param>
+        /// <param name="roll">The additional roll to apply</param>
         /// <param name="position">returned point on the spline, in world coords</param>
         /// <param name="rotation">returned rotation at the point on the spline, in world coords</param>
         /// <returns>True if the spline position is valid</returns>
         public static bool EvaluateSplineWithRoll(
             this CachedScaledSpline spline,
             Transform transform,
-            CinemachineSplineRoll roll,
-            Quaternion defaultRotation,
             float tNormalized, 
+            Quaternion defaultRotation,
+            CinemachineSplineRoll roll,
             out Vector3 position, out Quaternion rotation)
         {
-            var result = LocalEvaluateSplineWithRoll(spline, roll, defaultRotation, tNormalized, out position, out rotation);
+            var result = LocalEvaluateSplineWithRoll(spline, tNormalized, defaultRotation, roll, out position, out rotation);
             position = Matrix4x4.TRS(transform.position, transform.rotation, Vector3.one).MultiplyPoint3x4(position);
             rotation = transform.rotation * rotation;
             return result;
         }
 
+        /// <summary>Evaluate a spline's world position and rotation at a normalized spline index</summary>
+        /// <param name="spline">The spline in question</param>
+        /// <param name="tNormalized">The normalized position on the spline</param>
+        /// <param name="position">returned point on the spline, in world coords</param>
+        /// <returns>True if the spline position is valid</returns>
+        public static bool EvaluateSplinePosition(
+            this CachedScaledSpline spline, Transform transform, float tNormalized, out Vector3 position)
+        {
+            if (spline == null || !SplineUtility.Evaluate(spline, tNormalized, out var localPosition, out _, out _))
+            {
+                position = Vector3.zero;
+                return false;
+            }
+            position = Matrix4x4.TRS(transform.position, transform.rotation, Vector3.one).MultiplyPoint3x4(localPosition);
+            return true;
+        }
+        
         /// <summary>
         /// Get the maximum value for the spline position.  Minimum value is always 0.
         /// </summary>
@@ -102,7 +119,7 @@ namespace Unity.Cinemachine
         /// <param name="unit">The spline position is expressed in these units</param>
         /// This is needed because we don't have access to the spline's scale.</param>
         /// <returns></returns>
-        public static float GetMaxPosition(this CachedScaledSpline spline, PathIndexUnit unit)
+        public static float GetMaxPosition(this ISpline spline, PathIndexUnit unit)
         {
             switch (unit)
             {
@@ -125,9 +142,11 @@ namespace Unity.Cinemachine
         /// <param name="unit">The spline position is expressed in these units</param>
         /// This is needed because we don't have access to the spline's scale.</param>
         /// <returns>The clamped position value, respecting the specified units</returns>
-        public static float StandardizePosition(this CachedScaledSpline spline, float t, PathIndexUnit unit, out float maxPos)
+        public static float StandardizePosition(this ISpline spline, float t, PathIndexUnit unit, out float maxPos)
         {
             maxPos = spline.GetMaxPosition(unit);
+            if (float.IsNaN(t))
+                return 0;
             if (!spline.Closed)
                 return Mathf.Clamp(t, 0, maxPos);
             t %= maxPos;